Change the publication date convention
Originally, I was putting the date the post was written in
articles.lisp. I think this is misleading and confusing. I'm changing
the convention so that if I post old content, I put a note on the top
of the post explicitly stating when the content was written. The date
the post was published is used in articles.lisp.